main and config

+++ main.py
@@ -212,14 +212,17 @@ if __name__ == '__main__':
 
             df = pd.DataFrame(data)
 
-            _ = open(f'results/stats_repr_{dataset_name}_config_{config_name}.csv', 'w+')
-            _.close()
+            if not os.path.isfile(f'results/stats_repr_{dataset_name}_config_{config_name}.csv'):
 
-            stats = pd.read_csv(f'results/stats_repr_{dataset_name}_config_{config_name}.csv')
+                df.to_csv(f'results/stats_repr_{dataset_name}_config_{config_name}.csv', index = False)
 
-            stats = pd.concat([stats, df], ignore_index=True)
+            else:
 
-            stats.to_csv(f'results/stats_repr_{dataset_name}_config_{config_name}.csv', index = False)
+                stats = pd.read_csv(f'results/stats_repr_{dataset_name}_config_{config_name}.csv')
+
+                stats = pd.concat([stats, df], ignore_index=True)
+
+                stats.to_csv(f'results/stats_repr_{dataset_name}_config_{config_name}.csv', index = False)
 
         for _ in range(n_runs):
             trx_encoder_params = dict(
@@ -280,11 +283,15 @@ if __name__ == '__main__':
 
             df = pd.DataFrame(data)
 
-            _ = open(f'results/stats_contr_{dataset_name}_config_{config_name}.csv', 'w+')
-            _.close()
+            if not os.path.isfile(f'results/stats_contr_{dataset_name}_config_{config_name}.csv'):
 
-            stats = pd.read_csv(f'results/stats_contr_{dataset_name}_config_{config_name}.csv')
+                
+                df.to_csv(f'results/stats_contr_{dataset_name}_config_{config_name}.csv', index = False)
 
-            stats = pd.concat([stats, df], ignore_index=True)
+            else:
 
-            stats.to_csv(f'results/stats_contr_{dataset_name}_config_{config_name}.csv', index = False)
+                stats = pd.read_csv(f'results/stats_contr_{dataset_name}_config_{config_name}.csv')
+
+                stats = pd.concat([stats, df], ignore_index=True)
+
+                stats.to_csv(f'results/stats_contr_{dataset_name}_config_{config_name}.csv', index = False)
